+++ Tests/Editor/EditmodeTests.cs
@@ -1,19 +1,13 @@
-// using Beacon.Sdk.Beacon.Permission;
-// using NUnit.Framework;
-// using TezosSDK.Tezos;
-// using TezosSDK.Tezos.API.Models;
-//
-//
-// public class EditmodeTests
-// {
-//     // A Test behaves as an ordinary method
-//     [Test]
-//     public async void EditmodeTestsSimplePasses()
-//     {
-//         TezosConfig.Instance.Network = NetworkType.ghostnet;
-//         var tokenContract = new TokenContract();
-//         var txHash = await tokenContract.Deploy();
-//
-//         var a = 5;
-//     }
-// }
+using NUnit.Framework;
+
+namespace Tests.Editor
+{
+    public class EditmodeTests
+    {
+        // A Test behaves as an ordinary method
+        [Test]
+        public void EditmodeTestsSimplePasses()
+        {
+        }
+    }
+}