+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Missing namespaces
 
 ### Changed
-- Issue Report Form menu item path changed (new path: `Window/Tezos SDK for Unity/Report an Issue`)
+- Issue Report Form menu item path changed (new path: `Tools/Tezos SDK for Unity/Report an Issue`)
 
 
 ## [1.5.1] - 2023-06-27