+++ CHANGELOG.md
@@ -11,12 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Bug](https://github.com/trilitech/tezos-unity-sdk/issues/103) with closing Beacon database connections before app exit
 
 ### Added
-- FA2 token contract (deploy, mint and transfer entrypoints)
+- FA2 token contract (deploy, mint and transfer features)
 - Implemented IPFS and on-chain image uploaders for WebGL and Editor platforms
 
 ### Changed
 - Issue Report Form menu item path changed (new path: `Tools/Tezos SDK for Unity/Report an Issue`)
-- ITezosAPI renamed to ITezos refactored and divided in Wallet and API parts
+- ITezosAPI renamed to ITezos, refactored and divided in Wallet and API parts
+- Missing namespaces
 
 
 ## [1.5.1] - 2023-06-27
@@ -121,7 +122,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added auto releases with GH actions
 
 
-[unreleased]: https://github.com/trilitech/tezos-unity-sdk/compare/1.5.1...HEAD
+[unreleased]: https://github.com/trilitech/tezos-unity-sdk/compare/2.0.0...HEAD
+[2.0.0]: https://github.com/trilitech/tezos-unity-sdk/releases/tag/2.0.0
 [1.5.1]: https://github.com/trilitech/tezos-unity-sdk/releases/tag/1.5.1
 [1.5.0]: https://github.com/trilitech/tezos-unity-sdk/releases/tag/1.5.0
 [1.4.0]: https://github.com/trilitech/tezos-unity-sdk/releases/tag/1.4.0